stability: implement node liveness; first step towards new range leases
This change adds a node liveness table as a global system table.
Nodes periodically write updates to their liveness record by doing
a conditional put to the liveness table. The leader of the range
containing the node liveness table gossips the latest information
to the rest of the system.

Each node has a `NodeLiveness` object which can be used to query
the status of any other node to find out if it's live or non-live
according to the liveness threshold duration compared to the last
time it successfully heartbeat its liveness record.

The as-yet-unused `IncrementEpoch` mechanism is also added in this
PR, for eventual use with the planned epoch-based range leader leases.

Updated the range leader lease RFC to reflect current thinking.

+++ server/node.go
@@ -706,7 +706,7 @@ func (n *Node) writeSummaries(ctx context.Context) error {
 	if runErr := n.stopper.RunTask(func() {
 		nodeStatus := n.recorder.GetStatusSummary()
 		if nodeStatus != nil {
-			key := keys.NodeStatusKey(int32(nodeStatus.Desc.NodeID))
+			key := keys.NodeStatusKey(nodeStatus.Desc.NodeID)
 			// We use PutInline to store only a single version of the node
 			// status. There's not much point in keeping the historical
 			// versions as we keep all of the constituent data as
